feat: parse fvar with ttf-parser to reach 100% upstream parity
Switches macOS to use CoreText only to enumerate candidate font paths
(via CTFontManagerCopyAvailableFontURLs), then parses each file with
ttf-parser. CT's per-instance descriptors carry synthetic postscript
names that don't match upstream's family-derived prefix; this commit
rebuilds the face data from raw OS/2 + fvar bytes instead.

Key behaviour changes that close the diff:
- Iterate CT URLs deduped to unique file paths, then call
  CTFontManagerCreateFontDescriptorsFromURL on a fragment-less URL so
  hidden families inside TTC files (Noto Sans Zawgyi, Hiragino Kaku
  Gothic Pro/ProN) are visible.
- Build instance postscripts as `family.replace(' ', '') + '-' +
  style.replace(' ', '')`; ignore both the per-instance psNameID
  (sometimes lower-case CJK region codes like "cn" instead of "SC")
  and NameID 25 (often baked with the style, e.g.
  "STIXTwoTextItalic").
- Read raw OS/2.usWidthClass and clamp to 1..=9 to match upstream's
  treatment of the invalid value 0 (Pilgiche.ttf surfaces as
  stretch=1, not Normal=5).
- Truncate (cast) wght axis coords to u16 instead of rounding so
  Skia-Bold surfaces as weight=1 from wght=1.949997.
- Store axis values as f32 to match upstream's serialised precision
  (3.199997 vs 3.1999969482421875).
- Rewrite the en-US Subfamily "Oblique" -> "Italic" when an
  English-language sibling already spells the style "Italic"
  (Galvji-Oblique).
- Fall back to a CT-derived descriptor for bitmap-only files
  ttf-parser can't open, like NISC18030.ttf.
- Drop the user-shadows-system dedup; upstream keeps both copies.

routes::font_file now allows any path the discovery layer enumerated
rather than gating on cfg.font_dirs (which is empty by default on
macOS), so the binary endpoint can serve every listed font.

+++ src/fonts/mod.rs
@@ -7,7 +7,6 @@
 
 mod cache;
 mod dirs;
-#[cfg(not(target_os = "macos"))]
 mod parser;
 mod platform;
 
@@ -22,10 +21,13 @@ use std::sync::{Arc, OnceLock};
 pub struct AxisInfo {
     pub tag: String,
     pub name: String,
-    pub value: f64,
-    pub min: f64,
-    pub max: f64,
-    pub default: f64,
+    // Stored as f32 to match upstream's serialised precision: ttf-parser
+    // gives fixed16.16 fractions whose nearest f32 prints with ~7 digits
+    // (e.g. "3.199997"); f64 would print the full 17-digit f32 mantissa.
+    pub value: f32,
+    pub min: f32,
+    pub max: f32,
+    pub default: f32,
     pub hidden: bool,
 }
